Move get_all_mean_clue_frequencies_per_puzzle_size() to zebra_utils
